Odoo_Hackathon/Server/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
python-multipart
boto3
python-dotenv
